fix : 기존 쿠키 삭제 코드 수정(자동로그인에 쿠키 삭제 추가)

+++ run/src/main/java/com/guide/run/user/controller/SignController.java
@@ -154,7 +154,7 @@ public class SignController {
             for (Cookie cookie : cookies) {
                 if ("refreshToken".equals(cookie.getName())) {
                     // 잘못된 경로("/api/oauth/login")에 저장된 쿠키 삭제
-                    deleteOldRefreshTokenCookie(response);
+                    if(cookie.getPath().equals("/api/oauth/login")) deleteOldRefreshTokenCookie(response);
                     String refreshToken = cookie.getValue();
                     try {
